Begin conversion of dropdown, errors remaining
