--- packages/patternfly-4/react-core/src/components/Dropdown/DropdownGroup.d.ts
@@ -1,11 +1,0 @@
-import { FunctionComponent, HTMLProps, ReactNode } from 'react';
-import { Omit } from '../../helpers/typeUtils';
-
-export interface DropdownGroupProps extends Omit<React.HTMLProps<HTMLDivElement>, 'label'> {
-  children?: ReactNode;
-  label?: ReactNode;
-}
-
-declare const DropdownGroup: FunctionComponent<DropdownGroupProps>;
-
-export default DropdownGroup;
