+++ packages/patternfly-4/react-core/src/components/Dropdown/DropdownGroup.test.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import * as React from 'react';
 import { shallow } from 'enzyme';
-import DropdownGroup from './DropdownGroup';
+import { DropdownGroup } from './DropdownGroup';
 
 describe('dropdown groups', () => {
   test('basic render', () => {
