--- packages/patternfly-4/react-core/src/components/Dropdown/DropdownItem.d.ts
@@ -1,9 +1,0 @@
-import { FunctionComponent } from 'react';
-import { InternalDropdownItemProps } from './InternalDropdownItem';
-
-export interface DropdownItemProps extends InternalDropdownItemProps {
-}
-
-declare const DropdownItem: FunctionComponent<DropdownItemProps>;
-
-export default DropdownItem;
