+++ packages/patternfly-4/react-core/src/components/Dropdown/DropdownToggle.test.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import * as React from 'react';
 import { mount } from 'enzyme';
-import DropdownToggle from './DropdownToggle';
+import { DropdownToggle } from './DropdownToggle';
 
 describe('API', () => {
   test('click on closed', () => {
