+++ packages/patternfly-4/react-core/src/components/Dropdown/InternalDropdownItem.test.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import * as React from 'react';
 import { shallow } from 'enzyme';
-import InternalDropdownItem from './InternalDropdownItem';
-import DropdownSeparator from './Separator';
+import { InternalDropdownItem } from './InternalDropdownItem';
+import { DropdownSeparator } from './Separator';
 
 describe('dropdown items', () => {
   test('a', () => {
