+++ packages/patternfly-4/react-core/src/components/Dropdown/InternalDropdownItem.tsx
@@ -1,62 +1,58 @@
-import React from 'react';
+import * as React from 'react';
 import ReactDOM from 'react-dom';
-import PropTypes from 'prop-types';
 import { css } from '@patternfly/react-styles';
-import { componentShape } from '../../helpers/componentShape';
 import { DropdownContext } from './dropdownConstants';
 import { KEY_CODES, KEYHANDLER_DIRECTION } from '../../helpers/constants';
 import { Tooltip } from '../Tooltip';
 
-const propTypes = {
+export interface InternalDropdownItemProps extends React.HTMLProps<HTMLAnchorElement> {
   /** Anything which can be rendered as dropdown item */
-  children: PropTypes.node,
+  children?: React.ReactNode; 
   /** Classes applied to root element of dropdown item */
-  className: PropTypes.string,
+  className?: string; 
   /** Indicates which component will be used as dropdown item */
-  component: componentShape,
+  component?: React.ReactNode | string;
   /** Role for the item */
-  role: PropTypes.string,
+  role?: string; 
   /** Render dropdown item as disabled option */
-  isDisabled: PropTypes.bool,
+  isDisabled?: boolean; 
   /** Forces display of the hover state of the element */
-  isHovered: PropTypes.bool,
+  isHovered?: boolean; 
   /** Default hyperlink location */
-  href: PropTypes.string,
+  href?: string; 
   /** Tooltip to display when hovered over the item */
-  tooltip: PropTypes.node,
+  tooltip?: React.ReactNode;
   /** Additional tooltip props forwarded to the Tooltip component */
-  tooltipProps: PropTypes.any,
-  index: PropTypes.number,
-  context: PropTypes.shape({
-    keyHandler: PropTypes.func,
-    sendRef: PropTypes.func
-  }),
+  tooltipProps?: any;
+  index?: number; 
+  context?: {
+    keyHandler?: (index: number, direction: string) => void,
+    sendRef?: (index: number, ref: any, isDisabled: boolean) => void
+  };
   /** Callback for click event */
-  onClick: PropTypes.func,
-  /** Additional props are spread to the container component */
-  '': PropTypes.any // eslint-disable-line react/require-default-props
-};
+  onClick?: (event:React.MouseEvent<HTMLAnchorElement>|React.KeyboardEvent) => void;
+}
 
-const defaultProps = {
-  children: null,
-  className: '',
-  isHovered: false,
-  component: 'a',
-  role: null,
-  isDisabled: false,
-  href: '',
-  tooltip: null,
-  tooltipProps: {},
-  onClick: Function.prototype,
-  index: -1,
-  context: {
-    keyHandler: Function.prototype,
-    sendRef: Function.prototype
-  }
-};
-
-class DropdownItem extends React.Component {
+export class DropdownItem extends React.Component<InternalDropdownItemProps> {
   ref = React.createRef();
+
+  static defaultProps = {
+    children: null,
+    className: '',
+    isHovered: false,
+    component: 'a',
+    role: null,
+    isDisabled: false,
+    href: '',
+    tooltip: null,
+    tooltipProps: {},
+    onClick: Function.prototype,
+    index: -1,
+    context: {
+      keyHandler: Function.prototype,
+      sendRef: Function.prototype
+    }
+  };
 
   componentDidMount() {
     this.props.context.sendRef(this.props.index, this.ref.current, this.props.isDisabled);
@@ -169,8 +165,3 @@ class DropdownItem extends React.Component {
     );
   }
 }
-
-DropdownItem.propTypes = propTypes;
-DropdownItem.defaultProps = defaultProps;
-
-export default DropdownItem;
