--- packages/patternfly-4/react-core/src/components/Dropdown/KebabToggle.d.ts
@@ -1,9 +1,0 @@
-import { FunctionComponent } from 'react';
-import { DropdownToggleProps } from './DropdownToggle';
-
-export interface KebabProps extends DropdownToggleProps {
-}
-
-declare const Kebab: FunctionComponent<KebabProps>;
-
-export default Kebab;
