--- packages/patternfly-4/react-core/src/components/Dropdown/Separator.d.ts
@@ -1,8 +1,0 @@
-import { FunctionComponent } from 'react';
-
-export interface SeparatorProps extends React.HTMLProps<HTMLDivElement> {
-}
-
-declare const Separator: FunctionComponent<SeparatorProps>;
-
-export default Separator;
