+++ packages/patternfly-4/react-core/src/components/Dropdown/Separator.tsx
@@ -1,10 +1,17 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import * as React from 'react';
 import { css } from '@patternfly/react-styles';
 import { DropdownContext, DropdownArrowContext } from './dropdownConstants';
-import DropdownItem from './InternalDropdownItem';
+import { DropdownItem } from './InternalDropdownItem';
 
-const Separator = ({ className, ...props }) => (
+export interface SeparatorProps extends React.HTMLProps<HTMLDivElement> {
+ /** Classes applied to root element of dropdown item */
+ className?: string; 
+}
+
+export const Separator: React.Component<SeparatorProps> = ({
+  className = '',
+  ...props
+}: SeparatorProps) => (
   <DropdownContext.Consumer>
     {({ separatorClass }) => (
       <DropdownArrowContext.Consumer>
@@ -21,16 +28,3 @@ const Separator = ({ className, ...props }) => (
     )}
   </DropdownContext.Consumer>
 );
-
-Separator.propTypes = {
-  /** Classes applied to root element of dropdown item */
-  className: PropTypes.string,
-  /** Additional props are passed to the DropdownItem */
-  '': PropTypes.any // eslint-disable-line react/require-default-props
-};
-
-Separator.defaultProps = {
-  className: ''
-};
-
-export default Separator;
