+++ packages/patternfly-4/react-core/src/components/Dropdown/Toggle.test.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import * as React from 'react';
 import { mount } from 'enzyme';
-import DropdownToggle from './DropdownToggle';
-import KebabToggle from './KebabToggle';
+import { DropdownToggle } from './DropdownToggle';
+import { KebabToggle } from './KebabToggle';
 
 test('Dropdown toggle', () => {
   const view = mount(<DropdownToggle id="Dropdown Toggle">Dropdown</DropdownToggle>);
