--- packages/patternfly-4/react-core/src/components/Dropdown/dropdownConstants.d.ts
@@ -1,31 +1,0 @@
-import * as React from 'react';
-
-export enum DropdownPosition {
-  right = 'right',
-  left = 'left'
-}
-
-export enum DropdownDirection {
-  up = 'up',
-  down = 'down'
-}
-
-export const DropdownContext: React.Context<{
-  onSelect: (event: any) => void,
-  menuClass?: string,
-  itemClass?: string,
-  toggleClass?: string,
-  baseClass?: string,
-  baseComponent?: string,
-  sectionClass?: string,
-  sectionTitleClass?: string,
-  sectionComponent?: string,
-  disabledClass?: string,
-  hoverClass?: string,
-  separatorClass?: string
-}>
-
-export const DropdownArrowContext: React.Context<{
-  keyHandler: (index: number, position: any, custom: boolean) => void,
-  sendRef: (index: number, node: any, isDisabled: boolean) => void
-}>
