--- packages/patternfly-4/react-core/src/components/Dropdown/index.d.ts
@@ -1,8 +1,0 @@
-export { default as Dropdown, DropdownProps } from './Dropdown';
-export { DropdownDirection, DropdownPosition, DropdownContext } from './dropdownConstants';
-export { default as DropdownItem, DropdownItemProps } from './DropdownItem';
-export { default as DropdownSeparator, SeparatorProps } from './Separator';
-export { default as KebabToggle, KebabProps } from './KebabToggle';
-export { default as DropdownToggle, DropdownToggleProps } from './DropdownToggle';
-export { default as DropdownToggleCheckbox, DropdownToggleCheckboxProps } from './DropdownToggleCheckbox';
-export { default as DropdownGroup, DropdownGroupProps } from './DropdownGroup';
