--- packages/patternfly-4/react-core/src/components/Dropdown/index.js
@@ -1,8 +1,0 @@
-export { default as Dropdown } from './Dropdown';
-export { DropdownDirection, DropdownPosition, DropdownContext } from './dropdownConstants';
-export { default as DropdownGroup } from './DropdownGroup';
-export { default as DropdownItem } from './DropdownItem';
-export { default as DropdownSeparator } from './Separator';
-export { default as KebabToggle } from './KebabToggle';
-export { default as DropdownToggle } from './DropdownToggle';
-export { default as DropdownToggleCheckbox } from './DropdownToggleCheckbox';
