+++ packages/patternfly-4/react-core/src/components/Dropdown/index.ts
@@ -1,0 +1,8 @@
+export * from './Dropdown';
+export * from './dropdownConstants';
+export * from './DropdownGroup';
+export * from './DropdownItem';
+export * from './Separator';
+export * from './KebabToggle';
+export * from './DropdownToggle';
+export * from './DropdownToggleCheckbox';
